Using proper toolchain in purescript_rules.
We modify the whole purescript_rules sub-project to use a proper
toolchain instead of a set hardcoded hardcoded paths.

A side effect of this is to make this project both compatible with
"bindist-based" toolchain and "native" toolchains. In our specific case,
it makes possible to build this whole project using a regular linux.